make files and class names analogous btwn Programs & Procedures
