+++ qcengine/procedures/model.py
@@ -6,7 +6,7 @@ from pydantic import BaseModel
 from ..util import model_wrapper
 
 
-class BaseProcedure(BaseModel, abc.ABC):
+class ProcedureExecutor(BaseModel, abc.ABC):
 
     name: str
     procedure: str
