+++ qcengine/programs/cfour.py
@@ -6,7 +6,7 @@ from typing import Dict
 from qcelemental.util import which, safe_version
 
 from ..util import execute
-from .executor import ProgramExecutor
+from .model import ProgramExecutor
 
 pp = pprint.PrettyPrinter(width=120, compact=True, indent=1)
