+++ qcengine/programs/psi4.py
@@ -7,7 +7,7 @@ from typing import Dict
 from qcelemental.models import Result
 from qcelemental.util import parse_version, safe_version, which
 
-from .executor import ProgramExecutor
+from .model import ProgramExecutor
 from ..exceptions import InputError, RandomError, ResourceError, UnknownError
 from ..util import execute, popen
