+++ qcengine/programs/torchani.py
@@ -7,7 +7,7 @@ from typing import Dict
 from qcelemental.models import Provenance, Result
 from qcelemental.util import parse_version, safe_version, which_import
 
-from .executor import ProgramExecutor
+from .model import ProgramExecutor
 from ..exceptions import InputError, ResourceError
 from ..units import ureg
